// app/src/actions/types.ts
export enum types {
  APP_START = "APP_START",
  SET_IS_SOCKET_CONNECTED = "SET_IS_SOCKET_CONNECTED",
  SET_NAME = "SET_NAME",
  SET_MESSAGE = "SET_MESSAGE",
  SEND_MESSAGE = "SEND_MESSAGE",
  SET_MESSAGES = "SET_MESSAGES",
  SET_IS_MOUSE_DOWN = "SET_IS_MOUSE_DOWN",
  SET_GROUP_COUNT = "SET_GROUP_COUNT",
  SET_DRAWING_POINT = "SET_DRAWING_POINT",
  SET_DRAWING_POINTS = "SET_DRAWING_POINTS",
  SET_BROADCASTED_DRAWING_POINT = "SET_BROADCASTED_DRAWING_POINT",
  SET_BROADCASTED_DRAWING_POINTS = "SET_BROADCASTED_DRAWING_POINTS",
  SET_BROADCASTED_DRAWING_POINTS_GROUP = "SET_BROADCASTED_DRAWING_POINTS_GROUP",
  SET_POINTS_CACHE = "SET_POINTS_CACHE"
}